Regression Test Failure with 4441 Upload Error
Following the previous update, the system failed a regression test and reported a 4441 error while trying to submit a new Lab report.
This was due to the condition added on line 298 in Course.php and reversing this change solved the problem.

This condition was added to allow all Group members to be updated once the Tutor Marks the Lab Report but a better method to
archieve this has been implemented on line 367 in Course.php