Script.php: [bugfix] fix -- "Students unable to submit assignment after a recent change"
This commit temporarily fixed the bug reported at http://118.25.96.118/bugzilla/show_bug.cgi?id=65

This bug was introduced after a recent functionality addition, i.e., allowing Lecturer to edit assignment information.
See the following pull request for details.
https://github.com/lanlab-org/LRR/pull/21

It seems that $group_id is not properly assigned after merging the above pull request.
Therefore, when someone tried to insert a record to the database table lab_report_submissions, he encountered a running error.

In this fix, I just set $group_id 0 to avoid the database insertion error.
This is OK when an assignment is of type Individual.
Of course, this fix is not ideal, as it cannot handle the situation where the assignment is of type Group.
In the future, $group_id must be properly initialized, depending on the assignment type.

-Hui