Update signup.php
For the LRR website, we don't need to display both the sign-in and sign-up forms on one page. The best thing that we can do to have a clean and neat LRR website UI/UX we should only display the logo and sign in form. Then we are also going to put a signup link on the sign-in page that will lead to the signup form and on the signup form the top left corner lrr logo will lead us back to the sign-in form. 
On the sign-up form I have removed the portal for inputting the passport and added an if statement for inputting student number on the same form.