Update index.php
Updated the Logo from "LRRS" to "LRR".
Removed the signup form from the index.php file (Index.php file had two forms now with only one form for signing in and a link that leads to the signup form).
On the signup form, we have removed the portal for putting the passport because I have realized that our colleagues from this repository "Ashik-Nafis_Bug 53" only removed the passport portal from the form. 
We also added a background and some CSS properties for the hovers effects on the new form.